# api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aegis Trader
  description: AI-assisted, risk-gated automated trading platform. Paper trading only
    in V1; the LLM can only create trade proposals.
  version: 0.1.0
paths:
...
  /api/v1/markets/status:
    get:
      tags:
      - markets
      summary: Market session status
      operationId: market_status_api_v1_markets_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketStatusSchema'
      security:
      - HTTPBearer: []
  /api/v1/markets/overview:
    get:
      tags:
      - markets
      summary: Watchlist overview
      description: Configured symbol universe with quotes and the latest strategy
        signal.
      operationId: market_overview_api_v1_markets_overview_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarketOverviewSchema'
      security:
      - HTTPBearer: []
  /api/v1/markets/search:
    get:
      tags:
      - markets
      summary: Search assets
      operationId: search_assets_api_v1_markets_search_get
      security:
      - HTTPBearer: []
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          maxLength: 32
          description: Symbol or name fragment
          title: Q
        description: Symbol or name fragment
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 25
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssetSearchSchema'
                title: Response Search Assets Api V1 Markets Search Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/markets/{symbol}/quote:
    get:
      tags:
      - markets
      summary: Latest quote
      operationId: get_quote_api_v1_markets__symbol__quote_get
      security:
      - HTTPBearer: []
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
          title: Symbol
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/markets/{symbol}/candles:
    get:
      tags:
      - markets
      summary: Historical candles
      operationId: get_candles_api_v1_markets__symbol__candles_get
      security:
      - HTTPBearer: []
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
          title: Symbol
      - name: timeframe
        in: query
        required: false
        schema:
          type: string
          description: 'Candle timeframe: 1m, 5m, 15m, 30m, 1h, 4h, 1d, 1w'
          default: 1h
          title: Timeframe
        description: 'Candle timeframe: 1m, 5m, 15m, 30m, 1h, 4h, 1d, 1w'
      - name: start
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Inclusive start (ISO 8601)
          title: Start
        description: Inclusive start (ISO 8601)
      - name: end
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Inclusive end (ISO 8601)
          title: End
        description: Inclusive end (ISO 8601)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandleSeriesSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/markets/{symbol}/indicators:
    get:
      tags:
      - markets
      summary: Technical indicators
      operationId: get_indicators_api_v1_markets__symbol__indicators_get
      security:
      - HTTPBearer: []
      parameters:
      - name: symbol
        in: path
        required: true
        schema:
          type: string
          title: Symbol
      - name: timeframe
        in: query
        required: false
        schema:
          type: string
          description: 'Candle timeframe: 1m, 5m, 15m, 30m, 1h, 4h, 1d, 1w'
          default: 1h
          title: Timeframe
        description: 'Candle timeframe: 1m, 5m, 15m, 30m, 1h, 4h, 1d, 1w'
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndicatorResponseSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/proposals:
    post:
      tags:
      - proposals
      summary: Create Proposal
      operationId: create_proposal_api_v1_proposals_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - proposals
      summary: List Proposals
      operationId: list_proposals_api_v1_proposals_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ProposalStatus'
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalPageSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/proposals/{proposal_id}:
    get:
      tags:
      - proposals
      summary: Get Proposal
      operationId: get_proposal_api_v1_proposals__proposal_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: proposal_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Proposal Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalDetailSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/proposals/{proposal_id}/evaluate:
    post:
      tags:
      - proposals
      summary: Evaluate Proposal
      operationId: evaluate_proposal_api_v1_proposals__proposal_id__evaluate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: proposal_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Proposal Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalEvaluationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/proposals/{proposal_id}/execute:
    post:
      tags:
      - proposals
      summary: Execute Proposal
      operationId: execute_proposal_api_v1_proposals__proposal_id__execute_post
      security:
      - HTTPBearer: []
      parameters:
      - name: proposal_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Proposal Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionOutcome'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/proposals/{proposal_id}/cancel:
    post:
      tags:
      - proposals
      summary: Cancel Proposal
      operationId: cancel_proposal_api_v1_proposals__proposal_id__cancel_post
      security:
      - HTTPBearer: []
      parameters:
      - name: proposal_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Proposal Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AssetSearchSchema:
      properties:
        symbol:
          type: string
          title: Symbol
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        asset_class:
          type: string
          title: Asset Class
        exchange:
          anyOf:
          - type: string
          - type: 'null'
          title: Exchange
        currency:
          type: string
          title: Currency
        provider:
          type: string
          title: Provider
      type: object
      required:
      - symbol
      - name
      - asset_class
      - exchange
      - currency
      - provider
      title: AssetSearchSchema
...
    CandleSeriesSchema:
      properties:
        symbol:
          type: string
          title: Symbol
        timeframe:
          type: string
          title: Timeframe
        provider:
          type: string
          title: Provider
        candles:
          items:
            $ref: '#/components/schemas/CandleSchema'
          type: array
          title: Candles
        is_stale:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Stale
        age_seconds:
          anyOf:
          - type: number
          - type: 'null'
          title: Age Seconds
      type: object
      required:
      - symbol
      - timeframe
      - provider
      - candles
      title: CandleSeriesSchema
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    IndicatorResponseSchema:
      properties:
        symbol:
          anyOf:
          - type: string
          - type: 'null'
          title: Symbol
        timeframe:
          anyOf:
          - type: string
          - type: 'null'
          title: Timeframe
        price:
          anyOf:
          - type: number
          - type: 'null'
          title: Price
        data_timestamp:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Data Timestamp
        calculated_at:
          type: string
          format: date-time
          title: Calculated At
        age_seconds:
          anyOf:
          - type: number
          - type: 'null'
          title: Age Seconds
        is_stale:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Stale
        sma:
          $ref: '#/components/schemas/SeriesSchema'
        ema:
          $ref: '#/components/schemas/SeriesSchema'
        rsi:
          $ref: '#/components/schemas/SeriesSchema'
        macd:
          $ref: '#/components/schemas/MACDSchema'
        atr:
          $ref: '#/components/schemas/SeriesSchema'
        bollinger:
          $ref: '#/components/schemas/BollingerSchema'
        volume:
          $ref: '#/components/schemas/VolumeSchema'
      type: object
      required:
      - symbol
      - timeframe
      - price
      - data_timestamp
      - calculated_at
      - age_seconds
      - is_stale
      - sma
      - ema
      - rsi
      - macd
      - atr
      - bollinger
      - volume
      title: IndicatorResponseSchema
...
    MarketOverviewSchema:
      properties:
        status:
          $ref: '#/components/schemas/MarketStatusSchema'
        items:
          items:
            $ref: '#/components/schemas/MarketOverviewItemSchema'
          type: array
          title: Items
      type: object
      required:
      - status
      - items
      title: MarketOverviewSchema
...
    MarketStatusSchema:
      properties:
        market:
          type: string
          title: Market
        is_open:
          type: boolean
          title: Is Open
        session:
          $ref: '#/components/schemas/MarketSession'
        opens_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Opens At
        closes_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Closes At
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        provider:
          type: string
          title: Provider
      type: object
      required:
      - market
      - is_open
      - session
      - opens_at
      - closes_at
      - timestamp
      - provider
      title: MarketStatusSchema
...
    ProposalCreate:
      properties:
        symbol:
          type: string
          maxLength: 32
          minLength: 1
          title: Symbol
        side:
          $ref: '#/components/schemas/TradeSide'
        order_type:
          $ref: '#/components/schemas/OrderType'
          default: MARKET
        quantity:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Quantity
        notional:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Notional
        limit_price:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Limit Price
        stop_price:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Stop Price
        stop_loss:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Stop Loss
        take_profit:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Take Profit
        strategy_signal_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Strategy Signal Id
        time_horizon:
          $ref: '#/components/schemas/TimeHorizon'
          default: SWING
        confidence:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Confidence
          default: '0.5'
        reasoning_summary:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Reasoning Summary
        idempotency_key:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Idempotency Key
      type: object
      required:
      - symbol
      - side
      title: ProposalCreate
      description: Manual TradeProposal creation payload (no execution).
...
    ProposalEvaluationResponse:
      properties:
        proposal:
          $ref: '#/components/schemas/ProposalSchema'
        evaluation:
          $ref: '#/components/schemas/RiskEvaluationSchema'
      type: object
      required:
      - proposal
      - evaluation
      title: ProposalEvaluationResponse
...
    ProposalSchema:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        portfolio_id:
          type: string
          format: uuid
          title: Portfolio Id
        strategy_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Strategy Id
        strategy_signal_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Strategy Signal Id
        symbol:
          type: string
          title: Symbol
        asset_class:
          $ref: '#/components/schemas/AssetClass'
        action:
          $ref: '#/components/schemas/OrderAction'
        order_type:
          $ref: '#/components/schemas/OrderType'
        source:
          $ref: '#/components/schemas/ProposalSource'
        status:
          $ref: '#/components/schemas/ProposalStatus'
        proposed_quantity:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Proposed Quantity
        proposed_position_percentage:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Proposed Position Percentage
        requested_notional:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Requested Notional
        entry_price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Entry Price
        limit_price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Limit Price
        stop_price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Stop Price
        stop_loss:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Stop Loss
        take_profit:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Take Profit
        confidence:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Confidence
        time_horizon:
          $ref: '#/components/schemas/TimeHorizon'
        reasoning_summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Reasoning Summary
        market_regime:
          anyOf:
          - type: string
          - type: 'null'
          title: Market Regime
        failure_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Failure Reason
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        decided_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Decided At
        executed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Executed At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - portfolio_id
      - symbol
      - asset_class
      - action
      - order_type
      - source
      - status
      - proposed_quantity
      - confidence
      - time_horizon
      - created_at
      - updated_at
      title: ProposalSchema
...
    ProposalStatus:
      type: string
      enum:
      - DRAFT
      - PENDING
      - PENDING_RISK
      - RISK_APPROVED
      - RISK_REJECTED
      - READY_FOR_EXECUTION
      - EXECUTING
      - EXECUTED
      - APPROVED
      - REJECTED
      - EXPIRED
      - CANCELLED
      - FAILED
      title: ProposalStatus
    QuoteSchema:
      properties:
        symbol:
          type: string
          title: Symbol
        bid:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Bid
        ask:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Ask
        last:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Last
        open:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Open
        high:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: High
        low:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Low
        previous_close:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Previous Close
        volume:
          anyOf:
          - type: integer
          - type: 'null'
          title: Volume
        currency:
          type: string
          title: Currency
        provider:
          type: string
          title: Provider
        market_timestamp:
          type: string
          format: date-time
          title: Market Timestamp
        received_at:
          type: string
          format: date-time
          title: Received At
        age_seconds:
          type: number
          title: Age Seconds
        is_stale:
          type: boolean
          title: Is Stale
      type: object
      required:
      - symbol
      - bid
      - ask
      - last
      - open
      - high
      - low
      - previous_close
      - volume
      - currency
      - provider
      - market_timestamp
      - received_at
      - age_seconds
      - is_stale
      title: QuoteSchema
...
    TradeSide:
      type: string
      enum:
      - BUY
      - SELL
      title: TradeSide
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer